+++ src/main/java/net/geforcemods/securitycraft/inventory/KeycardHolderMenu.java
@@ -79,6 +79,6 @@ public class KeycardHolderMenu extends AbstractContainerMenu {
 
 	@Override
 	public boolean stillValid(Player player) {
-		return PlayerUtils.getItemStackFromAnyHand(player, SCContent.KEYCARD_HOLDER.get()) == keycardHolderInv.getContainerStack();
+		return !PlayerUtils.getItemStackFromAnyHand(player, SCContent.KEYCARD_HOLDER.get()).isEmpty();
 	}
 }