+++ src/main/java/net/geforcemods/securitycraft/screen/CameraMonitorScreen.java
@@ -3,7 +3,9 @@ package net.geforcemods.securitycraft.screen;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.items.CameraMonitorItem;
 import net.geforcemods.securitycraft.network.server.MountCamera;
+import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.minecraft.core.GlobalPos;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
 public class CameraMonitorScreen extends CameraSelectScreen {
@@ -23,5 +25,10 @@ public class CameraMonitorScreen extends CameraSelectScreen {
 	@Override
 	protected void unbindCamera(int camID) {
 		CameraMonitorItem.removeCameraOnClient(camID, stack.getTag());
+	}
+
+	@Override
+	public boolean stillValid(Player player) {
+		return !PlayerUtils.getItemStackFromAnyHand(player, stack.getItem()).isEmpty();
 	}
 }